# api-laravel/public/swagger-ui/swagger.yaml
openapi: 3.0.0
info:
  title: api-laravel
  description: Backend для мобильного приложения ЭБС Лань
  version: 1.0.0
servers:
  - url: https://api.landev.ru
    description: Тестовая среда
  - url: http://api-laravel.ebs.local
    description: Локальная среда
components:
  securitySchemes:
    authToken:
      type: apiKey
      in: header
      name: x-auth-token
  schemas:
    BookListObject:
      type: object
      properties:
        all_count:
          type: integer
          example: 1
        expired:
          type: string
          example: "2099-12-31"
        id:
          type: integer
          example: 183134
        book_private:
          type: integer
          example: 0
        description:
          type: string
          example: "В монографии представлены научно-исследовательский материал..."
        year:
          type: integer
          example: 2021
        book_expired:
          type: string
          example: "2099-12-31"
          nullable: true
        title:
          type: string
          example: "Ива в ландшафтной архитектуре. Перспективное использование видов рода Salix в озеленении населенных пунктов"
        author:
          type: string
          example: "Вергунова А. А., Сокольская О. Б."
        hasPdf:
          type: boolean
          example: true
        hasEpub:
          type: boolean
          example: true
        hasAudio:
          type: boolean
          example: false
        hasSyntex:
          type: boolean
          example: true
        synthesizer_editor:
          type: string
          example: ""
        synthesizer_quality:
          type: integer
          example: 1
        publisher:
          type: string
          example: "Издательство \"Лань\""
        expired_date:
          type: string
          example: null
          nullable: true
        access_id:
          type: integer
          example: null
          nullable: true
        active:
          type: boolean
          example: false
    BookCardObject:
      type: object
      properties:
        id:
          type: integer
          example: 183134
        title:
          type: string
          example: "Ива в ландшафтной архитектуре. Перспективное использование видов рода Salix в озеленении населенных пунктов"
        description:
          type: string
          example: "В монографии представлены научно-исследовательский материал..."
        isbn:
          type: string
          example: "978-5-8114-7810-1"
        edition:
          type: string
          example: ""
        pages:
          type: string
          example: "120"
        book_private:
          type: integer
          example: 0
        publisher__fk:
          type: integer
          example: 905
        synthesizer_quality:
          type: integer
          example: 1
        author:
          type: string
          example: "Вергунова А. А., Сокольская О. Б."
        year:
          type: integer
          example: 2021
        active:
          type: boolean
          example: false
        hasPdf:
          type: boolean
          example: true
        hasEpub:
          type: boolean
          example: true
        hasSyntex:
          type: boolean
          example: true
        hasAudio:
          type: boolean
          example: false
        publisher:
          type: string
          example: "Издательство \"Лань\""
        expired:
          type: string
          example: "2099-12-31"
        synthesizer_editor:
          type: string
          example: ""
        access_date:
          type: string
          example: "2024-04-22"
        expired_date:
          type: string
          example: null
          nullable: true
        cover:
          type: string
          example: "http://ebs.local/img/cover/book/183134.jpg"
        audio:
          type: boolean
          example: false
        size:
          type: integer
          example: 0
...
    DocumentKeyObject:
      type: object
      properties:
        key:
          type: array
          items:
            type: integer
          example: [ 246,330,399,321,291,387,285,369,300,393,429,387,285,363,234,219,291,279,363,396,387,204,378,276,333,417,306,327,327,258,411,192,384,366,387,288,213,423,372,234,285,288,192,414,21 ]
...
    SubscriberNameObject:
      type: object
      properties:
        subscriber_name:
          type: string
          example: "Тестовый подписчик"
...
  responses:
    EmptySuccessResponse:
      description: Success response with empty data
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 200
              error:
                type: string
                example: ""
              status:
                type: string
                example: "OK"
              type:
                type: string
                example: "none"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
...
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 401
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "Пользователь не авторизован"
                  code:
                    type: integer
                    example: 2
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    NotFoundResource:
      description: Not Found response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 404
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ресурс не найден"
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    BookListSuccessResponse:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 200
              error:
                type: string
                example: ""
              status:
                type: string
                example: "OK"
              type:
                type: string
                example: "object"
              data:
                type: object
                properties:
                  cnt:
                    type: integer
                    example: 42
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookListObject'
...
    BookCardSuccessResponse:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 200
              error:
                type: string
                example: ""
              status:
                type: string
                example: "OK"
              type:
                type: string
                example: "object"
              data:
                $ref: '#/components/schemas/BookCardObject'
...
    AuthSuccessResponse:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 200
              error:
                type: string
                example: ""
              status:
                type: string
                example: "OK"
              type:
                type: string
                example: "object"
              data:
                $ref: '#/components/schemas/UserObject'
    AuthRegistrationRequiredResponse:
      description: Registration Required
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 401
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "Требуется регистрация"
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    AuthGeneralErrorResponse:
      description: Unexpected error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 500
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "Авторизация не удалась"
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    RegisterGeneralErrorResponse:
      description: General error response for registration
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 500
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "Регистрация не удалась"
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    FirstNameAndLastNameRequiredResponse:
      description: First name and last name required
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 400
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Введите фамилию и имя'
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    RegistrationNotAllowedResponse:
      description: Registration not allowed
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 400
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Для данного типа регистрации необходимо зайти в сеть вашей библиотеки'
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    LogOutErrorResponse:
      description: Logout error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 500
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Разлогин приложения прошел с ошибкой!'
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    RecoverPasswordErrorResponse:
      description: Error response for password recovery
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 400
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "Запрос на воосстановление не отправлен"
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    IpNotFound:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 401
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ip не найденн"
                  code:
                    type: integer
                    example: 2
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    CannotDeleteGuestUserResponse:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 403
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Мы не можем нарушить закон гостеприимства и удалить гостя.'
                  code:
                    type: integer
                    example: 2
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items:
                  type: object
                maxItems: 0
    SubscriberNameSuccessResponse:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 200
              error:
                type: string
                example: ""
              status:
                type: string
                example: "OK"
              type:
                type: string
                example: "object"
              data:
                $ref: '#/components/schemas/SubscriberNameObject'
    CreateDocumentKeySuccessResponse:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 200
              error:
                type: string
                example: ""
              status:
                type: string
                example: "OK"
              type:
                type: string
                example: "object"
              data:
                $ref: '#/components/schemas/DocumentKeyObject'
    NoAccessToResourceOrNotAvailableForDownloadError:
      description: No access to resource or not available for download
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                properties:
                  code:
                    type: integer
                    example: 403
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Нет доступа к ресурсу"
                  status:
                    type: string
                    example: "error"
                  type:
                    type: string
                    example: "object"
                  data:
                    type: array
                    items: { }
                    maxItems: 0
              - type: object
                properties:
                  code:
                    type: integer
                    example: 403
                  error:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "Недоступно для скачивания, обратитесь в ЭБС Лань или воспользуйтесь web-версией"
                  status:
                    type: string
                    example: "error"
                  type:
                    type: string
                    example: "object"
                  data:
                    type: array
                    items: { }
                    maxItems: 0
    NotAvailableForDownloadError:
      description: Not Available For Download
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 403
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "Недоступно для скачивания, обратитесь в ЭБС Лань или воспользуйтесь web-версией"
              status:
                type: string
                example: "error"
              type:
                type: string
                example: "object"
              data:
                type: array
                items: { }
                maxItems: 0
    GetMetaSuccessResponse:
      description: Success response
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 200
              error:
                type: string
                example: ""
              status:
                type: string
                example: "OK"
              type:
                type: string
                example: "object"
              data:
                $ref: '#/components/schemas/DocumentMetaObject'
...
security:
  - authToken: [ ]
paths:
  /ebs/1.1/book:
    get:
      summary: Получить книгу по ID
      operationId: BookCard
      tags:
        - Книги
      security:
        - authToken: [ ]
      parameters:
        - in: query
          name: id
          schema:
            type: integer
            example: 183134
          description: ID книги
      responses:
        200:
          $ref: '#/components/responses/BookCardSuccessResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /ebs/1.1/books:
    get:
      summary: Получить список книг
      description: получить список книг
      operationId: BookList
      tags:
        - Книги
      security:
        - authToken: [ ]
      parameters:
        - in: query
          name: category
          schema:
            type: integer
            example: 1
          description: ID Категории
        - in: query
          name: subcategory
          schema:
            type: integer
            example: 2
          description: ID подкатегории
        - in: query
          name: publisher
          schema:
            type: integer
            example: 3
          description: ID издателя
        - in: query
          name: author
          schema:
            type: string
            example: "Александр Пушкин"
          description: Имя автора
        - in: query
          name: sort
          schema:
            type: string
            example: "title"
          description: Поле сортировки
        - in: query
          name: syntex
          schema:
            type: integer
            example: 1
          description: Синтезатор речи (1 - показывать только те книжки у которых есть невизуальное чтение)
        - in: query
          name: limit
          schema:
            type: integer
            example: 10
          description: Лимит записей
        - in: query
          name: offset
          schema:
            type: integer
            example: 0
          description: Смещение пагинатора
      responses:
        200:
          $ref: '#/components/responses/BookListSuccessResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
...
  /ebs/1.1/users/current:
    post:
      summary: Авторизоваться в аккаунт через Email
      operationId: LogInByEmail
      tags:
        - Пользователь
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "ivan.p@yandex.ru"
                password:
                  type: string
                  example: "password123"
      responses:
        200:
          $ref: '#/components/responses/AuthSuccessResponse'
        401:
          $ref: '#/components/responses/AuthRegistrationRequiredResponse'
        500:
          $ref: '#/components/responses/AuthGeneralErrorResponse'
    delete:
      summary: Logout из аккаунта
      operationId: LogOutSocial
      tags:
        - Пользователь
      responses:
        200:
          $ref: '#/components/responses/EmptySuccessResponse'
        500:
          $ref: '#/components/responses/LogOutErrorResponse'
  /ebs/1.1/users/current/social:
    post:
      summary: Авторизоваться в аккаунт через соц сети
      operationId: LogInBySocial
      tags:
        - Пользователь
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                network:
                  type: string
                  example: "apple"
                token_social:
                  type: string
                  example: "eyJhbGciOiJFUzI1NiIsImtpZCI6IjABCDEFG1234567890HIJKLMNOPQRS1234567TUVXYZ8901234ABCDEFGHIJKL56789MNO"
      responses:
        200:
          $ref: '#/components/responses/AuthSuccessResponse'
        401:
          $ref: '#/components/responses/AuthRegistrationRequiredResponse'
        500:
          $ref: '#/components/responses/AuthGeneralErrorResponse'
    delete:
      summary: Logout из аккаунта соц сетей
      operationId: LogOutSocial
      tags:
        - Пользователь
      responses:
        200:
          $ref: '#/components/responses/EmptySuccessResponse'
        500:
          $ref: '#/components/responses/LogOutErrorResponse'
  /ebs/1.1/users/register:
    post:
      summary: Регистрация аккаунта через Email
      operationId: RegisterByEmail
      tags:
        - Пользователь
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Ivan Ivanov Ivanovich"
                email:
                  type: string
                  example: "ivan@mail.com"
                password:
                  type: string
                  example: "12345678"
      responses:
        200:
          $ref: '#/components/responses/EmptySuccessResponse'
        400:
          $ref: '#/components/responses/FirstNameAndLastNameRequiredResponse'
        500:
          $ref: '#/components/responses/RegisterGeneralErrorResponse'
  /ebs/1.1/users/register/social:
    post:
      summary: Регистрация аккаунта через соц сети
      operationId: RegisterBySocial
      tags:
        - Пользователь
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lastName:
                  type: string
                  example: "Ivanov"
                firstName:
                  type: string
                  example: "Ivan"
                email:
                  type: string
                  example: "ivan@mail.com"
                password:
                  type: string
                  example: "12345678"
                network:
                  type: string
                  example: "apple"
                token_social:
                  type: string
                  example: "eyJhbGciOiJFUzI1NiIsImtpZCI6IjABCDEFG1234567890HIJKLMNOPQRS1234567TUVXYZ8901234ABCDEFGHIJKL56789MNO"
                inviteCode:
                  type: string
                  example: "stS7YOlxmGDl"
      responses:
        200:
          $ref: '#/components/responses/AuthSuccessResponse'
        405:
          $ref: '#/components/responses/RegistrationNotAllowedResponse'
        500:
          $ref: '#/components/responses/RegisterGeneralErrorResponse'
  /ebs/1.1/users/recovery:
    post:
      summary: Восстановить пароль
      operationId: RecoverPassword
      tags:
        - Пользователь
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "ivan.p@yandex.ru"
      responses:
        200:
          $ref: '#/components/responses/EmptySuccessResponse'
        400:
          $ref: '#/components/responses/RecoverPasswordErrorResponse'
  /ebs/1.1/users/check_ip:
    get:
      summary: Проверить IP подписчика
      operationId: CheckIp
      tags:
        - Пользователь
      responses:
        200:
          $ref: '#/components/responses/SubscriberNameSuccessResponse'
        401:
          $ref: '#/components/responses/IpNotFound'
  /ebs/1.1/users/check_invite:
    get:
      summary: Проверить валидность инвайта
      operationId: CheckInvite
      tags:
        - Пользователь
      security:
        - authToken: [ ]
      parameters:
        - in: query
          name: invite
          required: true
          schema:
            type: string
            example: 123456-123456
            minLength: 9
            maxLength: 15
          description: Инвайт код из библиотеки
      responses:
        200:
          $ref: '#/components/responses/EmptySuccessResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /ebs/1.1/users/social/delete:
    put:
      summary: Выполнить запрос на деактивацию текущего пользователя
      operationId: DeactivateUser
      tags:
        - Пользователь
      security:
        - authToken: [ ]
      responses:
        200:
          $ref: '#/components/responses/EmptySuccessResponse'
        403:
          $ref: '#/components/responses/CannotDeleteGuestUserResponse'
...
  /ebs/1.1/book/key:
    get:
      summary: Получить ключ книги по ID
      operationId: CreateBookKey
      tags:
        - Книги
      security:
        - authToken: [ ]
      parameters:
        - in: query
          name: id
          schema:
            type: integer
            example: 192702
          description: ID книги
      responses:
        200:
          $ref: '#/components/responses/CreateDocumentKeySuccessResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/NoAccessToResourceOrNotAvailableForDownloadError'
        404:
          $ref: '#/components/responses/NotFoundResource'
  /ebs/1.1/book/meta:
    get:
      summary: Получить метаданные книги по ID
      operationId: GetBookMeta
      tags:
        - Книги
      security:
        - authToken: [ ]
      parameters:
        - in: query
          name: id
          schema:
            type: integer
            example: 192702
          description: ID книги
      responses:
        200:
          $ref: '#/components/responses/GetMetaSuccessResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/NotAvailableForDownloadError'
        404:
          $ref: '#/components/responses/NotFoundResource'
...